feat: PinNoteFragmentも変更に合わせ修正
