+++ modules/features/user/src/main/res/layout/fragment_pin_note.xml
@@ -1,12 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout>
-    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-            android:orientation="vertical" android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/pin_notes_view"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent" />
-    </LinearLayout>
+<layout xmlns:android="http://schemas.android.com/apk/res/android">
+    <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/pin_notes_view"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent" />
 </layout>
